Cleanup pace stencils testing  (#84)
* move things outside translate testing

* fix metric term edge_e/w dimension

* allow optional mask in Quantity

* move quantity_wrap and fix util test

* fix util test

* quantity auto detects mask

* more cleanup

* address PR comments

* Update pace-util/pace/util/quantity.py


* Update pace-util/pace/util/quantity.py
