+++ driver/pace/driver/run.py
@@ -18,11 +18,10 @@ from fv3core.testing.translate_fvdynamics import init_dycore_state_from_serializ
 from fv3gfs.physics import PhysicsState
 from fv3gfs.physics.stencils.physics import Physics
 from pace.dsl.stencil import StencilFactory
-from pace.stencils.testing.grid import DampingCoefficients, DriverGridData, GridData
 from pace.stencils.update_atmos_state import UpdateAtmosphereState
 from pace.util import QuantityFactory
 from pace.util.constants import N_HALO_DEFAULT
-from pace.util.grid import MetricTerms
+from pace.util.grid import DampingCoefficients, DriverGridData, GridData, MetricTerms
 
 
 class Driver:
