+++ fv3core/fv3core/stencils/d_sw.py
@@ -25,7 +25,6 @@ from fv3core.stencils.xtp_u import advect_u_along_x
 from fv3core.stencils.ytp_v import advect_v_along_y
 from pace.dsl.stencil import StencilFactory
 from pace.dsl.typing import FloatField, FloatFieldIJ, FloatFieldK
-from pace.stencils.testing.grid import DampingCoefficients, GridData
 from pace.util import (
     X_DIM,
     X_INTERFACE_DIM,
@@ -34,6 +33,7 @@ from pace.util import (
     Z_DIM,
     Z_INTERFACE_DIM,
 )
+from pace.util.grid import DampingCoefficients, GridData
 
 
 dcon_threshold = 1e-5
