+++ fv3core/fv3core/stencils/dyn_core.py
@@ -1,4 +1,4 @@
-from typing import Dict, Union
+from typing import Dict, Sequence, Union
 
 from gt4py.gtscript import (
     __INLINED,
@@ -29,16 +29,10 @@ from fv3core.stencils.del2cubed import HyperdiffusionDamping
 from fv3core.stencils.pk3_halo import PK3Halo
 from fv3core.stencils.riem_solver3 import RiemannSolver3
 from fv3core.stencils.riem_solver_c import RiemannSolverC
-from pace.dsl.stencil import StencilFactory
+from pace.dsl.stencil import GridIndexing, StencilFactory
 from pace.dsl.typing import FloatField, FloatFieldIJ, FloatFieldK
-from pace.stencils.testing.grid import (
-    DampingCoefficients,
-    GridData,
-    GridIndexing,
-    axis_offsets,
-    quantity_wrap,
-)
 from pace.util import X_DIM, Y_DIM, Z_DIM, Z_INTERFACE_DIM
+from pace.util.grid import DampingCoefficients, GridData
 
 
 HUGE_R = 1.0e40
@@ -163,6 +157,17 @@ def get_nk_heat_dissipation(
     return nk_heat_dissipation
 
 
+def _quantity_wrap(storage, dims: Sequence[str], grid_indexing: GridIndexing):
+    origin, extent = grid_indexing.get_origin_domain(dims)
+    return pace.util.Quantity(
+        storage,
+        dims=dims,
+        units="unknown",
+        origin=origin,
+        extent=extent,
+    )
+
+
 def dyncore_temporaries(grid_indexing: GridIndexing, *, backend: str):
     tmps: Dict[str, Union[pace.util.Quantity, "FloatField"]] = {}
     utils.storage_dict(
@@ -193,16 +198,16 @@ def dyncore_temporaries(grid_indexing: GridIndexing, *, backend: str):
         grid_indexing.origin_compute(add=(-grid_indexing.n_halo, 0, 0)),
         backend=backend,
     )
-    tmps["heat_source"] = quantity_wrap(
+    tmps["heat_source"] = _quantity_wrap(
         tmps["heat_source"], [X_DIM, Y_DIM, Z_DIM], grid_indexing
     )
-    tmps["divgd"] = quantity_wrap(
+    tmps["divgd"] = _quantity_wrap(
         tmps["divgd"],
         dims=[fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM, fv3util.Z_DIM],
         grid_indexing=grid_indexing,
     )
     for name in ["gz", "pkc", "zh"]:
-        tmps[name] = quantity_wrap(
+        tmps[name] = _quantity_wrap(
             tmps[name],
             dims=[fv3util.X_DIM, fv3util.Y_DIM, fv3util.Z_INTERFACE_DIM],
             grid_indexing=grid_indexing,
@@ -459,8 +464,7 @@ class AcousticDynamics:
             origin=grid_indexing.origin_full(),
             domain=grid_indexing.domain_full(),
         )
-        ax_offsets_pe = axis_offsets(
-            grid_indexing,
+        ax_offsets_pe = grid_indexing.axis_offsets(
             grid_indexing.origin_full(),
             grid_indexing.domain_full(add=(0, 0, 1)),
         )
