+++ fv3core/fv3core/stencils/fv_dynamics.py
@@ -15,7 +15,7 @@ from fv3core.stencils.remapping import LagrangianToEulerian
 from pace.dsl.stencil import FrozenStencil, StencilFactory
 from pace.dsl.typing import FloatField, FloatFieldIJ, FloatFieldK
 from pace.stencils.c2l_ord import CubedToLatLon
-from pace.stencils.testing.grid import DampingCoefficients, GridData
+from pace.util.grid import DampingCoefficients, GridData
 from pace.util.halo_updater import HaloUpdater
