+++ fv3core/fv3core/stencils/nh_p_grad.py
@@ -4,8 +4,8 @@ import pace.dsl.gt4py_utils as utils
 from fv3core.stencils.a2b_ord4 import AGrid2BGridFourthOrder
 from pace.dsl.stencil import StencilFactory
 from pace.dsl.typing import FloatField, FloatFieldIJ
-from pace.stencils.testing.grid import GridData
 from pace.util import Z_INTERFACE_DIM
+from pace.util.grid import GridData
 
 
 def set_k0_and_calc_wk(
