+++ fv3core/fv3core/stencils/remapping.py
@@ -23,7 +23,6 @@ from fv3core.stencils.moist_cv import moist_pt_func, moist_pt_last_step
 from fv3core.stencils.saturation_adjustment import SatAdjust3d
 from pace.dsl.stencil import StencilFactory
 from pace.dsl.typing import FloatField, FloatFieldIJ, FloatFieldK
-from pace.stencils.testing.grid import axis_offsets
 
 
 CONSV_MIN = 0.001
@@ -357,8 +356,7 @@ class LagrangianToEulerian:
             grid_indexing.jec,
         )
 
-        ax_offsets_jextra = axis_offsets(
-            grid_indexing,
+        ax_offsets_jextra = grid_indexing.axis_offsets(
             grid_indexing.origin_compute(),
             grid_indexing.domain_compute(add=(0, 1, 0)),
         )
