+++ fv3gfs-physics/tests/savepoint/translate/translate_update_dwind_phys.py
@@ -1,9 +1,9 @@
 import numpy as np
 
 import pace.util
-from pace.stencils.testing.grid import DriverGridData
 from pace.stencils.testing.translate_physics import TranslatePhysicsFortranData2Py
 from pace.stencils.update_dwind_phys import AGrid2DGridPhysics
+from pace.util.grid import DriverGridData
 
 
 class TranslateUpdateDWindsPhys(TranslatePhysicsFortranData2Py):
