+++ pace-util/pace/util/constants.py
@@ -12,6 +12,7 @@ Y_DIMS = (Y_DIM, Y_INTERFACE_DIM)
 Z_DIMS = (Z_DIM, Z_INTERFACE_DIM)
 HORIZONTAL_DIMS = X_DIMS + Y_DIMS
 INTERFACE_DIMS = (X_INTERFACE_DIM, Y_INTERFACE_DIM, Z_INTERFACE_DIM)
+SPATIAL_DIMS = X_DIMS + Y_DIMS + Z_DIMS
 
 WEST = 0
 EAST = 1
