+++ pace-util/pace/util/grid/generation.py
@@ -1,6 +1,6 @@
 import functools
 
-import pace.util as fv3util
+from pace import util
 from pace.dsl.gt4py_utils import asarray, make_storage_data
 from pace.dsl.stencil import GridIndexing
 from pace.stencils.corners import (
@@ -67,8 +67,8 @@ class MetricTerms:
     def __init__(
         self,
         *,
-        quantity_factory: fv3util.QuantityFactory,
-        communicator: fv3util.Communicator,
+        quantity_factory: util.QuantityFactory,
+        communicator: util.Communicator,
         grid_type: int = 0,
     ):
         assert grid_type < 3
@@ -84,14 +84,15 @@ class MetricTerms:
                 self.LON_OR_LAT_DIM: 2,
                 self.TILE_DIM: 6,
                 self.CARTESIAN_DIM: 3,
+                util.X_DIM: 1,
             }
         )
         self._grid_indexing = GridIndexing.from_sizer_and_communicator(
             self.quantity_factory.sizer, self._comm
         )
         self._grid_dims = [
-            fv3util.X_INTERFACE_DIM,
-            fv3util.Y_INTERFACE_DIM,
+            util.X_INTERFACE_DIM,
+            util.Y_INTERFACE_DIM,
             self.LON_OR_LAT_DIM,
         ]
         self._grid = self.quantity_factory.zeros(
@@ -102,9 +103,9 @@ class MetricTerms:
         npx, npy, ndims = self._tile_partitioner.global_extent(self._grid)
         self._npx = npx
         self._npy = npy
-        self._npz = self.quantity_factory.sizer.get_extent(fv3util.Z_DIM)[0]
+        self._npz = self.quantity_factory.sizer.get_extent(util.Z_DIM)[0]
         self._agrid = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_DIM, self.LON_OR_LAT_DIM], "radians", dtype=float
+            [util.X_DIM, util.Y_DIM, self.LON_OR_LAT_DIM], "radians", dtype=float
         )
         self._np = self._grid.np
         self._dx = None
@@ -192,11 +193,11 @@ class MetricTerms:
         npx: int,
         npy: int,
         npz: int,
-        communicator: fv3util.Communicator,
+        communicator: util.Communicator,
         backend: str,
         grid_type: int = 0,
     ) -> "MetricTerms":
-        sizer = fv3util.SubtileGridSizer.from_tile_params(
+        sizer = util.SubtileGridSizer.from_tile_params(
             nx_tile=npx - 1,
             ny_tile=npy - 1,
             nz=npz,
@@ -208,7 +209,7 @@ class MetricTerms:
             },
             layout=communicator.partitioner.tile.layout,
         )
-        quantity_factory = fv3util.QuantityFactory.from_backend(sizer, backend=backend)
+        quantity_factory = util.QuantityFactory.from_backend(sizer, backend=backend)
         return cls(
             quantity_factory=quantity_factory,
             communicator=communicator,
@@ -243,7 +244,7 @@ class MetricTerms:
 
     @property
     def lon(self):
-        return fv3util.Quantity(
+        return util.Quantity(
             data=self.grid.data[:, :, 0],
             dims=self.grid.dims[0:2],
             units=self.grid.units,
@@ -251,8 +252,8 @@ class MetricTerms:
         )
 
     @property
-    def lat(self) -> fv3util.Quantity:
-        return fv3util.Quantity(
+    def lat(self) -> util.Quantity:
+        return util.Quantity(
             data=self.grid.data[:, :, 1],
             dims=self.grid.dims[0:2],
             units=self.grid.units,
@@ -260,8 +261,8 @@ class MetricTerms:
         )
 
     @property
-    def lon_agrid(self) -> fv3util.Quantity:
-        return fv3util.Quantity(
+    def lon_agrid(self) -> util.Quantity:
+        return util.Quantity(
             data=self.agrid.data[:, :, 0],
             dims=self.agrid.dims[0:2],
             units=self.agrid.units,
@@ -269,8 +270,8 @@ class MetricTerms:
         )
 
     @property
-    def lat_agrid(self) -> fv3util.Quantity:
-        return fv3util.Quantity(
+    def lat_agrid(self) -> util.Quantity:
+        return util.Quantity(
             data=self.agrid.data[:, :, 1],
             dims=self.agrid.dims[0:2],
             units=self.agrid.units,
@@ -278,7 +279,7 @@ class MetricTerms:
         )
 
     @property
-    def dx(self) -> fv3util.Quantity:
+    def dx(self) -> util.Quantity:
         """
         the distance between grid corners along the x-direction
         """
@@ -287,7 +288,7 @@ class MetricTerms:
         return self._dx
 
     @property
-    def dy(self) -> fv3util.Quantity:
+    def dy(self) -> util.Quantity:
         """
         the distance between grid corners along the y-direction
         """
@@ -296,7 +297,7 @@ class MetricTerms:
         return self._dy
 
     @property
-    def dxa(self) -> fv3util.Quantity:
+    def dxa(self) -> util.Quantity:
         """
         the with of each grid cell along the x-direction
         """
@@ -305,7 +306,7 @@ class MetricTerms:
         return self._dx_agrid
 
     @property
-    def dya(self) -> fv3util.Quantity:
+    def dya(self) -> util.Quantity:
         """
         the with of each grid cell along the y-direction
         """
@@ -314,7 +315,7 @@ class MetricTerms:
         return self._dy_agrid
 
     @property
-    def dxc(self) -> fv3util.Quantity:
+    def dxc(self) -> util.Quantity:
         """
         the distance between cell centers along the x-direction
         """
@@ -323,7 +324,7 @@ class MetricTerms:
         return self._dx_center
 
     @property
-    def dyc(self) -> fv3util.Quantity:
+    def dyc(self) -> util.Quantity:
         """
         the distance between cell centers along the y-direction
         """
@@ -332,7 +333,7 @@ class MetricTerms:
         return self._dy_center
 
     @property
-    def ak(self) -> fv3util.Quantity:
+    def ak(self) -> util.Quantity:
         """
         the ak coefficient used to calculate the pressure at a given k-level:
         pk = ak + (bk * ps)
@@ -347,7 +348,7 @@ class MetricTerms:
         return self._ak
 
     @property
-    def bk(self) -> fv3util.Quantity:
+    def bk(self) -> util.Quantity:
         """
         the bk coefficient used to calculate the pressure at a given k-level:
         pk = ak + (bk * ps)
@@ -364,7 +365,7 @@ class MetricTerms:
     # TODO: can ks and ptop just be derived from ak and bk instead of being returned
     # as part of _set_hybrid_pressure_coefficients?
     @property
-    def ks(self) -> fv3util.Quantity:
+    def ks(self) -> util.Quantity:
         """
         the number of pure-pressure layers at the top of the model
         also the level where model transitions from pure pressure to
@@ -380,7 +381,7 @@ class MetricTerms:
         return self._ks
 
     @property
-    def ptop(self) -> fv3util.Quantity:
+    def ptop(self) -> util.Quantity:
         """
         the pressure of the top of atmosphere level
         """
@@ -394,7 +395,7 @@ class MetricTerms:
         return self._ptop
 
     @property
-    def ec1(self) -> fv3util.Quantity:
+    def ec1(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the x-direation at the cell centers
@@ -405,7 +406,7 @@ class MetricTerms:
         return self._ec1
 
     @property
-    def ec2(self) -> fv3util.Quantity:
+    def ec2(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the y-direation at the cell centers
@@ -416,7 +417,7 @@ class MetricTerms:
         return self._ec2
 
     @property
-    def ew1(self) -> fv3util.Quantity:
+    def ew1(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the x-direation at the left/right cell edges
@@ -427,7 +428,7 @@ class MetricTerms:
         return self._ew1
 
     @property
-    def ew2(self) -> fv3util.Quantity:
+    def ew2(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the y-direation at the left/right cell edges
@@ -438,7 +439,7 @@ class MetricTerms:
         return self._ew2
 
     @property
-    def cos_sg1(self) -> fv3util.Quantity:
+    def cos_sg1(self) -> util.Quantity:
         """
         Cosine of the angle at point 1 of the 'supergrid' within each grid cell:
         9---4---8
@@ -452,7 +453,7 @@ class MetricTerms:
         return self._cos_sg1
 
     @property
-    def cos_sg2(self) -> fv3util.Quantity:
+    def cos_sg2(self) -> util.Quantity:
         """
         Cosine of the angle at point 2 of the 'supergrid' within each grid cell:
         9---4---8
@@ -466,7 +467,7 @@ class MetricTerms:
         return self._cos_sg2
 
     @property
-    def cos_sg3(self) -> fv3util.Quantity:
+    def cos_sg3(self) -> util.Quantity:
         """
         Cosine of the angle at point 3 of the 'supergrid' within each grid cell:
         9---4---8
@@ -480,7 +481,7 @@ class MetricTerms:
         return self._cos_sg3
 
     @property
-    def cos_sg4(self) -> fv3util.Quantity:
+    def cos_sg4(self) -> util.Quantity:
         """
         Cosine of the angle at point 4 of the 'supergrid' within each grid cell:
         9---4---8
@@ -494,7 +495,7 @@ class MetricTerms:
         return self._cos_sg4
 
     @property
-    def cos_sg5(self) -> fv3util.Quantity:
+    def cos_sg5(self) -> util.Quantity:
         """
         Cosine of the angle at point 5 of the 'supergrid' within each grid cell:
         9---4---8
@@ -509,7 +510,7 @@ class MetricTerms:
         return self._cos_sg5
 
     @property
-    def cos_sg6(self) -> fv3util.Quantity:
+    def cos_sg6(self) -> util.Quantity:
         """
         Cosine of the angle at point 6 of the 'supergrid' within each grid cell:
         9---4---8
@@ -523,7 +524,7 @@ class MetricTerms:
         return self._cos_sg6
 
     @property
-    def cos_sg7(self) -> fv3util.Quantity:
+    def cos_sg7(self) -> util.Quantity:
         """
         Cosine of the angle at point 7 of the 'supergrid' within each grid cell:
         9---4---8
@@ -537,7 +538,7 @@ class MetricTerms:
         return self._cos_sg7
 
     @property
-    def cos_sg8(self) -> fv3util.Quantity:
+    def cos_sg8(self) -> util.Quantity:
         """
         Cosine of the angle at point 8 of the 'supergrid' within each grid cell:
         9---4---8
@@ -551,7 +552,7 @@ class MetricTerms:
         return self._cos_sg8
 
     @property
-    def cos_sg9(self) -> fv3util.Quantity:
+    def cos_sg9(self) -> util.Quantity:
         """
         Cosine of the angle at point 9 of the 'supergrid' within each grid cell:
         9---4---8
@@ -565,7 +566,7 @@ class MetricTerms:
         return self._cos_sg9
 
     @property
-    def sin_sg1(self) -> fv3util.Quantity:
+    def sin_sg1(self) -> util.Quantity:
         """
         Sine of the angle at point 1 of the 'supergrid' within each grid cell:
         9---4---8
@@ -579,7 +580,7 @@ class MetricTerms:
         return self._sin_sg1
 
     @property
-    def sin_sg2(self) -> fv3util.Quantity:
+    def sin_sg2(self) -> util.Quantity:
         """
         Sine of the angle at point 2 of the 'supergrid' within each grid cell:
         9---4---8
@@ -593,7 +594,7 @@ class MetricTerms:
         return self._sin_sg2
 
     @property
-    def sin_sg3(self) -> fv3util.Quantity:
+    def sin_sg3(self) -> util.Quantity:
         """
         Sine of the angle at point 3 of the 'supergrid' within each grid cell:
         9---4---8
@@ -607,7 +608,7 @@ class MetricTerms:
         return self._sin_sg3
 
     @property
-    def sin_sg4(self) -> fv3util.Quantity:
+    def sin_sg4(self) -> util.Quantity:
         """
         Sine of the angle at point 4 of the 'supergrid' within each grid cell:
         9---4---8
@@ -621,7 +622,7 @@ class MetricTerms:
         return self._sin_sg4
 
     @property
-    def sin_sg5(self) -> fv3util.Quantity:
+    def sin_sg5(self) -> util.Quantity:
         """
         Sine of the angle at point 5 of the 'supergrid' within each grid cell:
         9---4---8
@@ -636,7 +637,7 @@ class MetricTerms:
         return self._sin_sg5
 
     @property
-    def sin_sg6(self) -> fv3util.Quantity:
+    def sin_sg6(self) -> util.Quantity:
         """
         Sine of the angle at point 6 of the 'supergrid' within each grid cell:
         9---4---8
@@ -650,7 +651,7 @@ class MetricTerms:
         return self._sin_sg6
 
     @property
-    def sin_sg7(self) -> fv3util.Quantity:
+    def sin_sg7(self) -> util.Quantity:
         """
         Sine of the angle at point 7 of the 'supergrid' within each grid cell:
         9---4---8
@@ -664,7 +665,7 @@ class MetricTerms:
         return self._sin_sg7
 
     @property
-    def sin_sg8(self) -> fv3util.Quantity:
+    def sin_sg8(self) -> util.Quantity:
         """
         Sine of the angle at point 8 of the 'supergrid' within each grid cell:
         9---4---8
@@ -678,7 +679,7 @@ class MetricTerms:
         return self._sin_sg8
 
     @property
-    def sin_sg9(self) -> fv3util.Quantity:
+    def sin_sg9(self) -> util.Quantity:
         """
         Sine of the angle at point 9 of the 'supergrid' within each grid cell:
         9---4---8
@@ -692,7 +693,7 @@ class MetricTerms:
         return self._sin_sg9
 
     @property
-    def cosa(self) -> fv3util.Quantity:
+    def cosa(self) -> util.Quantity:
         """
         cosine of angle between coordinate lines at the cell corners
         averaged to ensure consistent answers
@@ -702,7 +703,7 @@ class MetricTerms:
         return self._cosa
 
     @property
-    def sina(self) -> fv3util.Quantity:
+    def sina(self) -> util.Quantity:
         """
         as cosa but sine
         """
@@ -711,7 +712,7 @@ class MetricTerms:
         return self._sina
 
     @property
-    def cosa_u(self) -> fv3util.Quantity:
+    def cosa_u(self) -> util.Quantity:
         """
         as cosa but defined at the left and right cell edges
         """
@@ -720,7 +721,7 @@ class MetricTerms:
         return self._cosa_u
 
     @property
-    def cosa_v(self) -> fv3util.Quantity:
+    def cosa_v(self) -> util.Quantity:
         """
         as cosa but defined at the top and bottom cell edges
         """
@@ -729,7 +730,7 @@ class MetricTerms:
         return self._cosa_v
 
     @property
-    def cosa_s(self) -> fv3util.Quantity:
+    def cosa_s(self) -> util.Quantity:
         """
         as cosa but defined at cell centers
         """
@@ -738,7 +739,7 @@ class MetricTerms:
         return self._cosa_s
 
     @property
-    def sina_u(self) -> fv3util.Quantity:
+    def sina_u(self) -> util.Quantity:
         """
         as cosa_u but with sine
         """
@@ -747,7 +748,7 @@ class MetricTerms:
         return self._sina_u
 
     @property
-    def sina_v(self) -> fv3util.Quantity:
+    def sina_v(self) -> util.Quantity:
         """
         as cosa_v but with sine
         """
@@ -756,7 +757,7 @@ class MetricTerms:
         return self._sina_v
 
     @property
-    def rsin_u(self) -> fv3util.Quantity:
+    def rsin_u(self) -> util.Quantity:
         """
         1/sina_u**2,
         defined as the inverse-squrared as it is only used as such
@@ -766,7 +767,7 @@ class MetricTerms:
         return self._rsin_u
 
     @property
-    def rsin_v(self) -> fv3util.Quantity:
+    def rsin_v(self) -> util.Quantity:
         """
         1/sina_v**2,
         defined as the inverse-squrared as it is only used as such
@@ -776,7 +777,7 @@ class MetricTerms:
         return self._rsin_v
 
     @property
-    def rsina(self) -> fv3util.Quantity:
+    def rsina(self) -> util.Quantity:
         """
         1/sina**2,
         defined as the inverse-squrared as it is only used as such
@@ -786,7 +787,7 @@ class MetricTerms:
         return self._rsina
 
     @property
-    def rsin2(self) -> fv3util.Quantity:
+    def rsin2(self) -> util.Quantity:
         """
         1/sin_sg5**2,
         defined as the inverse-squrared as it is only used as such
@@ -796,7 +797,7 @@ class MetricTerms:
         return self._rsin2
 
     @property
-    def l2c_v(self) -> fv3util.Quantity:
+    def l2c_v(self) -> util.Quantity:
         """
         angular momentum correction for converting v-winds
         from lat/lon to cartesian coordinates
@@ -806,7 +807,7 @@ class MetricTerms:
         return self._l2c_v
 
     @property
-    def l2c_u(self) -> fv3util.Quantity:
+    def l2c_u(self) -> util.Quantity:
         """
         angular momentum correction for converting u-winds
         from lat/lon to cartesian coordinates
@@ -816,7 +817,7 @@ class MetricTerms:
         return self._l2c_u
 
     @property
-    def es1(self) -> fv3util.Quantity:
+    def es1(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the x-direation at the top/bottom cell edges,
@@ -827,7 +828,7 @@ class MetricTerms:
         return self._es1
 
     @property
-    def es2(self) -> fv3util.Quantity:
+    def es2(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the y-direation at the top/bottom cell edges,
@@ -838,7 +839,7 @@ class MetricTerms:
         return self._es2
 
     @property
-    def ee1(self) -> fv3util.Quantity:
+    def ee1(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the x-direation at the cell corners,
@@ -849,7 +850,7 @@ class MetricTerms:
         return self._ee1
 
     @property
-    def ee2(self) -> fv3util.Quantity:
+    def ee2(self) -> util.Quantity:
         """
         cartesian components of the local unit vetcor
         in the y-direation at the cell corners,
@@ -860,7 +861,7 @@ class MetricTerms:
         return self._ee2
 
     @property
-    def divg_u(self) -> fv3util.Quantity:
+    def divg_u(self) -> util.Quantity:
         """
         sina_v * dyc/dx
         """
@@ -874,7 +875,7 @@ class MetricTerms:
         return self._divg_u
 
     @property
-    def divg_v(self) -> fv3util.Quantity:
+    def divg_v(self) -> util.Quantity:
         """
         sina_u * dxc/dy
         """
@@ -888,7 +889,7 @@ class MetricTerms:
         return self._divg_v
 
     @property
-    def del6_u(self) -> fv3util.Quantity:
+    def del6_u(self) -> util.Quantity:
         """
         sina_v * dx/dyc
         """
@@ -902,7 +903,7 @@ class MetricTerms:
         return self._del6_u
 
     @property
-    def del6_v(self) -> fv3util.Quantity:
+    def del6_v(self) -> util.Quantity:
         """
         sina_u * dy/dxc
         """
@@ -916,7 +917,7 @@ class MetricTerms:
         return self._del6_v
 
     @property
-    def vlon(self) -> fv3util.Quantity:
+    def vlon(self) -> util.Quantity:
         """
         unit vector in eastward longitude direction,
         3d array whose last dimension is length 3 and indicates x/y/z value
@@ -926,7 +927,7 @@ class MetricTerms:
         return self._vlon
 
     @property
-    def vlat(self) -> fv3util.Quantity:
+    def vlat(self) -> util.Quantity:
         """
         unit vector in northward latitude direction,
         3d array whose last dimension is length 3 and indicates x/y/z value
@@ -936,7 +937,7 @@ class MetricTerms:
         return self._vlat
 
     @property
-    def z11(self) -> fv3util.Quantity:
+    def z11(self) -> util.Quantity:
         """
         vector product of horizontal component of the cell-center vector
         with the unit longitude vector
@@ -946,7 +947,7 @@ class MetricTerms:
         return self._z11
 
     @property
-    def z12(self) -> fv3util.Quantity:
+    def z12(self) -> util.Quantity:
         """
         vector product of horizontal component of the cell-center vector
         with the unit latitude vector
@@ -956,7 +957,7 @@ class MetricTerms:
         return self._z12
 
     @property
-    def z21(self) -> fv3util.Quantity:
+    def z21(self) -> util.Quantity:
         """
         vector product of vertical component of the cell-center vector
         with the unit longitude vector
@@ -966,7 +967,7 @@ class MetricTerms:
         return self._z21
 
     @property
-    def z22(self) -> fv3util.Quantity:
+    def z22(self) -> util.Quantity:
         """
         vector product of vertical component of the cell-center vector
         with the unit latitude vector
@@ -976,7 +977,7 @@ class MetricTerms:
         return self._z22
 
     @property
-    def a11(self) -> fv3util.Quantity:
+    def a11(self) -> util.Quantity:
         """
         0.5*z22/sin_sg5
         """
@@ -985,7 +986,7 @@ class MetricTerms:
         return self._a11
 
     @property
-    def a12(self) -> fv3util.Quantity:
+    def a12(self) -> util.Quantity:
         """
         0.5*z21/sin_sg5
         """
@@ -994,7 +995,7 @@ class MetricTerms:
         return self._a12
 
     @property
-    def a21(self) -> fv3util.Quantity:
+    def a21(self) -> util.Quantity:
         """
         0.5*z12/sin_sg5
         """
@@ -1003,7 +1004,7 @@ class MetricTerms:
         return self._a21
 
     @property
-    def a22(self) -> fv3util.Quantity:
+    def a22(self) -> util.Quantity:
         """
         0.5*z11/sin_sg5
         """
@@ -1012,7 +1013,7 @@ class MetricTerms:
         return self._a22
 
     @property
-    def edge_w(self) -> fv3util.Quantity:
+    def edge_w(self) -> util.Quantity:
         """
         factor to interpolate scalars from a to c grid at the western grid edge
         """
@@ -1026,7 +1027,7 @@ class MetricTerms:
         return self._edge_w
 
     @property
-    def edge_e(self) -> fv3util.Quantity:
+    def edge_e(self) -> util.Quantity:
         """
         factor to interpolate scalars from a to c grid at the eastern grid edge
         """
@@ -1040,7 +1041,7 @@ class MetricTerms:
         return self._edge_e
 
     @property
-    def edge_s(self) -> fv3util.Quantity:
+    def edge_s(self) -> util.Quantity:
         """
         factor to interpolate scalars from a to c grid at the southern grid edge
         """
@@ -1054,7 +1055,7 @@ class MetricTerms:
         return self._edge_s
 
     @property
-    def edge_n(self) -> fv3util.Quantity:
+    def edge_n(self) -> util.Quantity:
         """
         factor to interpolate scalars from a to c grid at the northern grid edge
         """
@@ -1068,7 +1069,7 @@ class MetricTerms:
         return self._edge_n
 
     @property
-    def edge_vect_w(self) -> fv3util.Quantity:
+    def edge_vect_w(self) -> util.Quantity:
         """
         factor to interpolate vectors from a to c grid at the western grid edge
         """
@@ -1082,7 +1083,7 @@ class MetricTerms:
         return self._edge_vect_w
 
     @property
-    def edge_vect_e(self) -> fv3util.Quantity:
+    def edge_vect_e(self) -> util.Quantity:
         """
         factor to interpolate vectors from a to c grid at the eastern grid edge
         """
@@ -1096,7 +1097,7 @@ class MetricTerms:
         return self._edge_vect_e
 
     @property
-    def edge_vect_s(self) -> fv3util.Quantity:
+    def edge_vect_s(self) -> util.Quantity:
         """
         factor to interpolate vectors from a to c grid at the southern grid edge
         """
@@ -1110,7 +1111,7 @@ class MetricTerms:
         return self._edge_vect_s
 
     @property
-    def edge_vect_n(self) -> fv3util.Quantity:
+    def edge_vect_n(self) -> util.Quantity:
         """
         factor to interpolate vectors from a to c grid at the northern grid edge
         """
@@ -1124,7 +1125,7 @@ class MetricTerms:
         return self._edge_vect_n
 
     @property
-    def da_min(self) -> fv3util.Quantity:
+    def da_min(self) -> util.Quantity:
         """
         the minimum agrid cell area across all ranks,
         if mpi is not present and the communicator is a DummyComm this will be
@@ -1135,7 +1136,7 @@ class MetricTerms:
         return self._da_min
 
     @property
-    def da_max(self) -> fv3util.Quantity:
+    def da_max(self) -> util.Quantity:
         """
         the maximum agrid cell area across all ranks,
         if mpi is not present and the communicator is a DummyComm this will be
@@ -1146,7 +1147,7 @@ class MetricTerms:
         return self._da_max
 
     @property
-    def da_min_c(self) -> fv3util.Quantity:
+    def da_min_c(self) -> util.Quantity:
         """
         the minimum cgrid cell area across all ranks,
         if mpi is not present and the communicator is a DummyComm this will be
@@ -1157,7 +1158,7 @@ class MetricTerms:
         return self._da_min_c
 
     @property
-    def da_max_c(self) -> fv3util.Quantity:
+    def da_max_c(self) -> util.Quantity:
         """
         the maximum cgrid cell area across all ranks,
         if mpi is not present and the communicator is a DummyComm this will be
@@ -1168,21 +1169,21 @@ class MetricTerms:
         return self._da_max_c
 
     @cached_property
-    def area(self) -> fv3util.Quantity:
+    def area(self) -> util.Quantity:
         """
         the area of each a-grid cell
         """
         return self._compute_area()
 
     @cached_property
-    def area_c(self) -> fv3util.Quantity:
+    def area_c(self) -> util.Quantity:
         """
         the area of each c-grid cell
         """
         return self._compute_area_c()
 
     @cached_property
-    def _dgrid_xyz(self) -> fv3util.Quantity:
+    def _dgrid_xyz(self) -> util.Quantity:
         """
         cartesian coordinates of each dgrid cell center
         """
@@ -1191,7 +1192,7 @@ class MetricTerms:
         )
 
     @cached_property
-    def _agrid_xyz(self) -> fv3util.Quantity:
+    def _agrid_xyz(self) -> util.Quantity:
         """
         cartesian coordinates of each agrid cell center
         """
@@ -1202,11 +1203,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rarea(self) -> fv3util.Quantity:
+    def rarea(self) -> util.Quantity:
         """
         1/cell area
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.area.data,
             dims=self.area.dims,
             units="m^-2",
@@ -1214,11 +1215,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rarea_c(self) -> fv3util.Quantity:
+    def rarea_c(self) -> util.Quantity:
         """
         1/cgrid cell area
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.area_c.data,
             dims=self.area_c.dims,
             units="m^-2",
@@ -1226,11 +1227,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rdx(self) -> fv3util.Quantity:
+    def rdx(self) -> util.Quantity:
         """
         1/dx
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.dx.data,
             dims=self.dx.dims,
             units="m^-1",
@@ -1238,11 +1239,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rdy(self) -> fv3util.Quantity:
+    def rdy(self) -> util.Quantity:
         """
         1/dy
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.dy.data,
             dims=self.dy.dims,
             units="m^-1",
@@ -1250,11 +1251,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rdxa(self) -> fv3util.Quantity:
+    def rdxa(self) -> util.Quantity:
         """
         1/dxa
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.dxa.data,
             dims=self.dxa.dims,
             units="m^-1",
@@ -1262,11 +1263,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rdya(self) -> fv3util.Quantity:
+    def rdya(self) -> util.Quantity:
         """
         1/dya
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.dya.data,
             dims=self.dya.dims,
             units="m^-1",
@@ -1274,11 +1275,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rdxc(self) -> fv3util.Quantity:
+    def rdxc(self) -> util.Quantity:
         """
         1/dxc
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.dxc.data,
             dims=self.dxc.dims,
             units="m^-1",
@@ -1286,11 +1287,11 @@ class MetricTerms:
         )
 
     @cached_property
-    def rdyc(self) -> fv3util.Quantity:
+    def rdyc(self) -> util.Quantity:
         """
         1/dyc
         """
-        return fv3util.Quantity(
+        return util.Quantity(
             data=1.0 / self.dyc.data,
             dims=self.dyc.dims,
             units="m^-1",
@@ -1465,7 +1466,7 @@ class MetricTerms:
         )
 
     def _compute_dxdy(self):
-        dx = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], "m")
+        dx = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "m")
 
         dx.view[:, :] = great_circle_distance_along_axis(
             self._grid.view[:, :, 0],
@@ -1474,7 +1475,7 @@ class MetricTerms:
             self._np,
             axis=0,
         )
-        dy = self.quantity_factory.zeros([fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], "m")
+        dy = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "m")
         dy.view[:, :] = great_circle_distance_along_axis(
             self._grid.view[:, :, 0],
             self._grid.view[:, :, 1],
@@ -1500,8 +1501,8 @@ class MetricTerms:
 
     def _compute_dxdy_agrid(self):
 
-        dx_agrid = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "m")
-        dy_agrid = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "m")
+        dx_agrid = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "m")
+        dy_agrid = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "m")
         lon, lat = self._grid.data[:, :, 0], self._grid.data[:, :, 1]
         lon_y_center, lat_y_center = lon_lat_midpoint(
             lon[:, :-1], lon[:, 1:], lat[:, :-1], lat[:, 1:], self._np
@@ -1535,12 +1536,8 @@ class MetricTerms:
         return dx_agrid, dy_agrid
 
     def _compute_dxdy_center(self):
-        dx_center = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], "m"
-        )
-        dy_center = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], "m"
-        )
+        dx_center = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "m")
+        dy_center = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "m")
 
         lon_agrid, lat_agrid = (
             self._agrid.data[:-1, :-1, 0],
@@ -1599,7 +1596,7 @@ class MetricTerms:
         return dx_center, dy_center
 
     def _compute_area(self):
-        area = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "m^2")
+        area = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "m^2")
         area.data[:, :] = -1.0e8
 
         area.data[3:-4, 3:-4] = get_area(
@@ -1613,7 +1610,7 @@ class MetricTerms:
 
     def _compute_area_c(self):
         area_cgrid = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], "m^2"
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], "m^2"
         )
         area_cgrid.data[3:-3, 3:-3] = get_area(
             self._agrid.data[2:-3, 2:-3, 0],
@@ -1655,8 +1652,8 @@ class MetricTerms:
     def _set_hybrid_pressure_coefficients(self):
         ks = self.quantity_factory.zeros([], "")
         ptop = self.quantity_factory.zeros([], "mb")
-        ak = self.quantity_factory.zeros([fv3util.Z_INTERFACE_DIM], "mb")
-        bk = self.quantity_factory.zeros([fv3util.Z_INTERFACE_DIM], "")
+        ak = self.quantity_factory.zeros([util.Z_INTERFACE_DIM], "mb")
+        bk = self.quantity_factory.zeros([util.Z_INTERFACE_DIM], "")
         pressure_coefficients = set_hybrid_pressure_coefficients(self._npz)
         ks = pressure_coefficients.ks
         ptop = pressure_coefficients.ptop
@@ -1666,10 +1663,10 @@ class MetricTerms:
 
     def _calculate_center_vectors(self):
         ec1 = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_DIM, self.CARTESIAN_DIM], ""
+            [util.X_DIM, util.Y_DIM, self.CARTESIAN_DIM], ""
         )
         ec2 = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_DIM, self.CARTESIAN_DIM], ""
+            [util.X_DIM, util.Y_DIM, self.CARTESIAN_DIM], ""
         )
         ec1.data[:] = self._np.nan
         ec2.data[:] = self._np.nan
@@ -1685,10 +1682,10 @@ class MetricTerms:
 
     def _calculate_vectors_west(self):
         ew1 = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM, self.CARTESIAN_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM, self.CARTESIAN_DIM], ""
         )
         ew2 = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM, self.CARTESIAN_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM, self.CARTESIAN_DIM], ""
         )
         ew1.data[:] = self._np.nan
         ew2.data[:] = self._np.nan
@@ -1705,10 +1702,10 @@ class MetricTerms:
 
     def _calculate_vectors_south(self):
         es1 = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
         )
         es2 = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
         )
         es1.data[:] = self._np.nan
         es2.data[:] = self._np.nan
@@ -1724,34 +1721,22 @@ class MetricTerms:
         return es1, es2
 
     def _calculate_more_trig_terms(self, cos_sg, sin_sg):
-        cosa_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
-        cosa_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
-        cosa_s = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        sina_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
-        sina_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
-        rsin_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
-        rsin_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
+        cosa_u = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
+        cosa_v = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        cosa_s = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        sina_u = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
+        sina_v = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        rsin_u = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
+        rsin_v = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
         rsina = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
-        rsin2 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        rsin2 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         cosa = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
         sina = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
         (
             cosa.data[:, :],
@@ -1791,33 +1776,33 @@ class MetricTerms:
     def _init_cell_trigonometry(self):
 
         self._cosa_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM], ""
         )
         self._cosa_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM], ""
         )
-        self._cosa_s = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        self._cosa_s = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         self._sina_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM], ""
         )
         self._sina_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM], ""
         )
         self._rsin_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM], ""
         )
         self._rsin_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM], ""
         )
         self._rsina = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
-        self._rsin2 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        self._rsin2 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         self._cosa = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
         self._sina = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
 
         # This section calculates the cos_sg and sin_sg terms, which describe the
@@ -1869,11 +1854,11 @@ class MetricTerms:
         supergrid_trig = {}
         for i in range(1, 10):
             supergrid_trig[f"cos_sg{i}"] = self.quantity_factory.zeros(
-                [fv3util.X_DIM, fv3util.Y_DIM], ""
+                [util.X_DIM, util.Y_DIM], ""
             )
             supergrid_trig[f"cos_sg{i}"].data[:-1, :-1] = cos_sg[:, :, i - 1]
             supergrid_trig[f"sin_sg{i}"] = self.quantity_factory.zeros(
-                [fv3util.X_DIM, fv3util.Y_DIM], ""
+                [util.X_DIM, util.Y_DIM], ""
             )
             supergrid_trig[f"sin_sg{i}"].data[:-1, :-1] = sin_sg[:, :, i - 1]
 
@@ -1902,33 +1887,33 @@ class MetricTerms:
         in-place without the halo updates. For use only in validation tests.
         """
         self._cosa_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM], ""
         )
         self._cosa_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM], ""
         )
-        self._cosa_s = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        self._cosa_s = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         self._sina_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM], ""
         )
         self._sina_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM], ""
         )
         self._rsin_u = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_DIM], ""
         )
         self._rsin_v = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_DIM, util.Y_INTERFACE_DIM], ""
         )
         self._rsina = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
-        self._rsin2 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        self._rsin2 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         self._cosa = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
         self._sina = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM], ""
         )
 
         cos_sg = self._np.array(
@@ -1981,12 +1966,8 @@ class MetricTerms:
         )
 
     def _calculate_latlon_momentum_correction(self):
-        l2c_v = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
-        l2c_u = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
+        l2c_v = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
+        l2c_u = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
         (
             l2c_v.data[self._halo : -self._halo, self._halo : -self._halo - 1],
             l2c_u.data[self._halo : -self._halo - 1, self._halo : -self._halo],
@@ -1995,10 +1976,10 @@ class MetricTerms:
 
     def _calculate_xy_unit_vectors(self):
         ee1 = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
         )
         ee2 = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
+            [util.X_INTERFACE_DIM, util.Y_INTERFACE_DIM, self.CARTESIAN_DIM], ""
         )
         ee1.data[:] = self._np.nan
         ee2.data[:] = self._np.nan
@@ -2011,18 +1992,10 @@ class MetricTerms:
         return ee1, ee2
 
     def _calculate_divg_del6(self):
-        del6_u = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
-        del6_v = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
-        divg_u = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
-        divg_v = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
+        del6_u = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        del6_v = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
+        divg_u = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        divg_v = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
         sin_sg = [
             self.sin_sg1.data[:-1, :-1],
             self.sin_sg2.data[:-1, :-1],
@@ -2064,18 +2037,10 @@ class MetricTerms:
         As _calculate_divg_del6 but updates self.divg and self.del6 attributes
         in-place without the halo updates. For use only in validation tests.
         """
-        del6_u = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
-        del6_v = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
-        divg_u = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
-        )
-        divg_v = self.quantity_factory.zeros(
-            [fv3util.X_INTERFACE_DIM, fv3util.Y_DIM], ""
-        )
+        del6_u = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        del6_v = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
+        divg_u = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        divg_v = self.quantity_factory.zeros([util.X_INTERFACE_DIM, util.Y_DIM], "")
         sin_sg = [
             self.sin_sg1.data[:-1, :-1],
             self.sin_sg2.data[:-1, :-1],
@@ -2108,10 +2073,10 @@ class MetricTerms:
 
     def _calculate_unit_vectors_lonlat(self):
         vlon = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_DIM, self.CARTESIAN_DIM], ""
+            [util.X_DIM, util.Y_DIM, self.CARTESIAN_DIM], ""
         )
         vlat = self.quantity_factory.zeros(
-            [fv3util.X_DIM, fv3util.Y_DIM, self.CARTESIAN_DIM], ""
+            [util.X_DIM, util.Y_DIM, self.CARTESIAN_DIM], ""
         )
 
         vlon.data[:-1, :-1], vlat.data[:-1, :-1] = unit_vector_lonlat(
@@ -2120,10 +2085,10 @@ class MetricTerms:
         return vlon, vlat
 
     def _calculate_grid_z(self):
-        z11 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        z12 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        z21 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        z22 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        z11 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        z12 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        z21 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        z22 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         (
             z11.data[:-1, :-1],
             z12.data[:-1, :-1],
@@ -2139,10 +2104,10 @@ class MetricTerms:
         return z11, z12, z21, z22
 
     def _calculate_grid_a(self):
-        a11 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        a12 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        a21 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
-        a22 = self.quantity_factory.zeros([fv3util.X_DIM, fv3util.Y_DIM], "")
+        a11 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        a12 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        a21 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
+        a22 = self.quantity_factory.zeros([util.X_DIM, util.Y_DIM], "")
         (
             a11.data[:-1, :-1],
             a12.data[:-1, :-1],
@@ -2159,13 +2124,13 @@ class MetricTerms:
 
     def _calculate_edge_factors(self):
         nhalo = self._halo
-        edge_s = self.quantity_factory.zeros([fv3util.X_INTERFACE_DIM], "")
-        edge_n = self.quantity_factory.zeros([fv3util.X_INTERFACE_DIM], "")
-        edge_e = self.quantity_factory.zeros([fv3util.Y_INTERFACE_DIM], "")
-        edge_w = self.quantity_factory.zeros([fv3util.Y_INTERFACE_DIM], "")
+        edge_s = self.quantity_factory.zeros([util.X_INTERFACE_DIM], "")
+        edge_n = self.quantity_factory.zeros([util.X_INTERFACE_DIM], "")
+        edge_e = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
+        edge_w = self.quantity_factory.zeros([util.X_DIM, util.Y_INTERFACE_DIM], "")
         (
-            edge_w.data[nhalo:-nhalo],
-            edge_e.data[nhalo:-nhalo],
+            edge_w.data[:, nhalo:-nhalo],
+            edge_e.data[:, nhalo:-nhalo],
             edge_s.data[nhalo:-nhalo],
             edge_n.data[nhalo:-nhalo],
         ) = edge_factors(
@@ -2181,10 +2146,10 @@ class MetricTerms:
         return edge_w, edge_e, edge_s, edge_n
 
     def _calculate_edge_a2c_vect_factors(self):
-        edge_vect_s = self.quantity_factory.zeros([fv3util.X_DIM], "")
-        edge_vect_n = self.quantity_factory.zeros([fv3util.X_DIM], "")
-        edge_vect_e = self.quantity_factory.zeros([fv3util.Y_DIM], "")
-        edge_vect_w = self.quantity_factory.zeros([fv3util.Y_DIM], "")
+        edge_vect_s = self.quantity_factory.zeros([util.X_DIM], "")
+        edge_vect_n = self.quantity_factory.zeros([util.X_DIM], "")
+        edge_vect_e = self.quantity_factory.zeros([util.Y_DIM], "")
+        edge_vect_w = self.quantity_factory.zeros([util.Y_DIM], "")
         (
             edge_vect_w.data[:-1],
             edge_vect_e.data[:-1],
@@ -2218,7 +2183,7 @@ class MetricTerms:
             self._da_min_c = min_area_c
             self._da_max_c = max_area_c
 
-    def split_cartesian_into_storages(self, var: fv3util.Quantity):
+    def split_cartesian_into_storages(self, var: util.Quantity):
         """
         Provided a quantity of dims [X_DIM, Y_DIM, CARTESIAN_DIM]
              or [X_INTERFACE_DIM, Y_INTERFACE_DIM, CARTESIAN_DIM]
