+++ pace-util/pace/util/grid/geometry.py
@@ -652,7 +652,7 @@ def edge_factors(
                 grid, agrid, nhalo, radius, istart, iend, np
             )
 
-    return edge_w, edge_e, edge_s, edge_n
+    return edge_w[np.newaxis, :], edge_e[np.newaxis, :], edge_s, edge_n
 
 
 def set_west_edge_factor(grid, agrid, nhalo, radius, jstart, jend, np):
